Renamed the Ardrey Sprint0 Commands namespace to Sprintfinity.Commands.
