+++ Sprintfinity3902/Commands/SetLinkRightCommand.cs
@@ -1,18 +1,19 @@
-﻿using Sprintfinity3902.Interfaces;
+﻿using Ardrey.Sprint0;
+using Sprintfinity3902.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Ardrey.Sprint0.Commands
+namespace Sprintfinity3902.Commands
 {
-    public class SetLinkRightCommand : ICommand 
+    public class SetLinkRightCommand : ICommand
     {
         Player PlayerCharacter;
 
         public SetLinkRightCommand(Player player)
         {
             PlayerCharacter = player;
-            
+
         }
 
         public void Execute()
