+++ Sprintfinity3902/Commands/SetLinkUpCommand.cs
@@ -1,9 +1,10 @@
-﻿using Sprintfinity3902.Interfaces;
+﻿using Ardrey.Sprint0;
+using Sprintfinity3902.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Ardrey.Sprint0.Commands
+namespace Sprintfinity3902.Commands
 {
     public class SetLinkUpCommand : ICommand
     {
